Algo de vector esta cagando todo, posiblemente haya que cambiarlo porque es una cagada

+++ ant.h
@@ -10,8 +10,8 @@
 #define NEURON_RANGE 10
 
 typedef struct {
-	float home;
-	float food;
+	int home;
+	int food;
 }pheromones;
 
 class ant
